config introduced
config will be used to configure its neighboring client package.

GCSCli config will be populated using the fields of the
directory deployment manifest when integrated with bosh-cli,
hence the json tags for consistency with the rest of the
manifest.
github.com/cloudfoundry/bosh-cli

As this is exposed to user input, we perform significant
validation. This ensures that the user can find badness
as soon as possible when deploying.

Authentication is currently exposed as an explicit
Service Account File or implicitly as Application Default
Credentials. The GCS client package provides more options
but these should cover most cases until users start
making specific requests.

This follows the configuration and testing style used in s3cli.
github.com/pivotal-golang/s3cli

Change-Id: I8de7191a160306796a5e86bd6acce1a6e8d0dfb9