cli introduced + integration test structure built
cli is the primary interface used by the bosh director to
access the blobstore. This is effectively the main.go file
from s3cli slightly modified for gcs naming. The same interface
as s3cli is exposed, allowing integration tests to be directly
adapated with minimal gcs-specific tweaks.
github.com/pivotal-golang/s3cli
https://github.com/cloudfoundry/bosh

The initial version of the integration tests covers the cli's
basic lifecycle under a single configuration. This will be
expanded in a future change to avoid an oversized commit.
This reuses as much code from s3cli's integration tests as
possible.

Change-Id: I45f5b993745ba5945cf3964f052114a0647bec26