config switched location handling semantics
The previous working assumption was that gcscli would create
buckets as necessary; this did not match up with s3cli, which
requires pre-existing buckets.

We switched to validating the desired StorageClass against the
pre-existing bucket's location. This ensures that configuration
errors happen immediately rather than when the first Put
is attempted.

This also means we use both Regional and MultiRegional buckets
for integration testing.

All location names were switched to upper-case to match the
Location value filled in a BucketAttrs when the metadata is
queried from the blobstore.
https://godoc.org/cloud.google.com/go/storage#BucketAttrs

Change-Id: I51f2182cd09e6e6ea4f9d6891627f01161b82daa
Reviewed-on: https://team-review.git.corp.google.com/74089
Reviewed-by: Jeff Johnson <jrjohnson@google.com>