cli clarified usage
This includes description and examples adapted from the README.

Previously, using the tool without the README for reference would
be difficult, this makes it self contained.

Now, -h or -help prints a full description of the flags as well
as some usage examples. Not providing a configuration file gives
the user a recommendation to call -help.

closes #2

Change-Id: I97f2942599d84dde10845eb68836b06d7cb6f57b
Reviewed-on: https://team-review.git.corp.google.com/74092
Reviewed-by: Jeff Johnson <jrjohnson@google.com>