config + client introduced customer-supplied encryption keys
GCS provides Data Encryption for content at rest.
Additionally, there is the option for customers to supply their
own encryption keys. As bosh can store sensitive data in its
blobstore, some customers may want to protect this data
using keys they own.

This includes an encryption_key field in the configuration.
We validate that field to either be empty or exactly the length
allowed of a GCS key to avoid runtime errors.

This includes associated integration tests. There is duplication
of the encryption-related TableEntries. If we end up with a third
set of duplicated entries, I'll refactor to have the bucket names
set in the BeforeSuite rather than individual suites.

closes #5

Change-Id: Ic0d5e181501ec8cb490bbeef114de4381529e1f8
Reviewed-on: https://team-review.git.corp.google.com/74668
Reviewed-by: Jeff Johnson <jrjohnson@google.com>