ci introduced custom Dockerfile
Previously, we used a pre-existing bosh-google-cpi-release docker
image for CI. That image had dependencies we did not need and
could change in incompatible ways.
https://github.com/cloudfoundry-incubator/bosh-google-cpi-release/

This contains the Dockerfile defining our minimal set of
dependencies as well as a set of system_tests. These tests are
treated as unit tests to check environment sanity.
The Dockerfile is used to populate boshcpi/bosh-gcscli
https://hub.docker.com/r/boshcpi/bosh-gcscli/

closes #8
closes #6

Change-Id: If2df74053f1e0fa674e2dc80dde8b81e897f846d
Reviewed-on: https://team-review.git.corp.google.com/75710
Reviewed-by: Meaghan Kjelland <mkjelland@google.com>