config introduced 'none' credentials_source
This allows users to indicate read-only access to the desired
bucket. This is necessary for bosh-director as it may wish to
access third party blobs which are public without needing to
specify credentials.
https://github.com/cloudfoundry/bosh

The exact error for using a read-only client in a mutating
operation is reused from s3cli
https://github.com/pivotal-golang/s3cli/

This change includes corresponding unit and integration tests.
A new set of integration tests, gcs_public_test, is added which
runs against a publicly available bucket.

Makefile now manages public bucket to accomodate the new test set.
As it is explicitly required to be publicly accessible when tests
are run, we need to wait for IAM to propagate to the endpoint
used by the GCS SDK.

closes #9

Change-Id: I4698386e11723cb5955baa42b950d1863c2b6468
Reviewed-on: https://team-review.git.corp.google.com/75892
Reviewed-by: Jeff Johnson <jrjohnson@google.com>