config introduced credential_source new semantics.
This introduces the service_account_file property on the
configuration and brings credentials_source inline with s3cli
as being a flag for determining operation rather than being
a path to a file on disk.

This will be followed up with an additional change
to implement the actual client functionality.

Change-Id: I512482b5bbe6cc0a095454523f7ac2ed162b544e
Reviewed-on: https://team-review.git.corp.google.com/77222
Reviewed-by: Meaghan Kjelland <mkjelland@google.com>