ci explicitly set gcloud project
An assumption was that the project associated with the service
account would be used. However, this is not the case.
If the concourse host is configured with a GCP project,
at least the project configuration leaks through into
the concourse task.

This corrects that by adding the 'google_project' parameteter to
all integration tests.

Change-Id: I3b9dedfa4ad2d76ac0f4de21110b3388afb1f336
Reviewed-on: https://team-review.git.corp.google.com/79128
Reviewed-by: Meaghan Kjelland <mkjelland@google.com>