config/client: default to read-only client
Default to a read-only client if Application Default Credentials fail to load.
This issue is notable on GCE where grabbing the DefaultTokenSource will
throw an error if the machine doesn't have a service account with valid
scope. In that scenario we don't want to fail public blobstore access.

This change removes a public API and updates an existing one. Consumers
should make the following change:

-  gcsClient, err := client.NewSDK(ctx, gcsConfig)
-  if err != nil {
-     log.Fatalf("creating gcs sdk: %v\n", err)
-  }
-  blobstoreClient, err := client.New(ctx, gcsClient, &gcsConfig)
+  blobstoreClient, err := client.New(ctx, &gcsConfig)