client: try with public client, then authenticated
The pervious commit tried to special case the credentials source for
GCE. this is a partial solution. what if a GCE user wanted to provide
credentials via a known path?
  eg ~/.config/gcloud/application_default_credentials.json

This commit changes the behavior of Get/Exists to first try with a
public client and if that fails to use the authenticated client.

Changes:
- update newStorageClient(s) to return a public GCS client and an
  authenticated one if avaliable
- update Get/Exists to try with both clients if they exist, Public
  preferred.

fixes #17