Replace the deprecated golint with go vet
golint has been deprecated and is now recommending the usage of either
staticcheck or go vet.

Because of the way golint was previously installed, it was adding it to
the go.mod file, causing problems when running unit-tests. Since golint
is deprecated, it's best to replace it instead.

go vet is not 100% equivalent, but can detect some problems the linter
caught in the past.

[#180054765] TAS-138 Add Golang autobumping and dependency autobumping to our Go repos