Updates error checking for library version bump
Version 1.51.0 of the 'cloud.google.com/go/storage' library wraps
NotFound errors for buckets and objects. As described in the PR
introducing the change, this breaks client code that doesn't use
'errors.Is()' to check the error type.

This commit updates the client code to work the way the library now
expects it to.

The PR that introduces the breaking change in question is here:
<https://github.com/googleapis/google-cloud-go/pull/11519>